requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
elasticsearch[async]==8.17.0
chromadb==0.5.23
temporalio==1.9.0
httpx==0.28.1
boto3==1.36.1
python-magic==0.4.27
pytesseract==0.3.13
Pillow==11.1.0
unstructured==0.16.11
python-docx==1.1.2
openpyxl==3.1.5
sentence-transformers==3.3.1
pydantic==2.10.4
pydantic-settings==2.7.1
pytest==8.3.4
pytest-asyncio==0.25.0
aiofiles==24.1.0
